[wip] kbs_protocol: support multiple attesters at once
The AA and kbs_protocol are designed to only use one attester to gather
evidence. This commit makes attesters not mutually exclusive. As many
attesters as are detected are loaded and the evidence is a combination
of what they provide.

To support runtime attestation, or more specifically, the EventLog
struct, we identify just one attester that supports runtime attestation
(if the option is selected). This allows us to avoid extra complexity in
the EventLog logic while keeping our attesters independent and
non-hierarchical.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>