attesters: add device class to attesters and evidence
To help keep track of multiple devices, let's add an additional
way for attesters to identify themselves.

Specifically, allow each attester to specify a device class string.
This will be used for a few different things. Mainly, it will end up
as the name of the submodule in the EAR token.
This is central to how multiple devices will be exposed to the resource
policy and ultimately the user.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>